+++ src/main/java/net/Mega2223/AguaEngine/graphics/objects/shapes/Line.java
@@ -1,7 +1,7 @@
 package net.Mega2223.AguaEngine.graphics.objects.shapes;
 
 public class Line extends Shape {
-    public float[][] points = {};
+    public double[][] points = {};
     public Line(){
         this.type = SHAPE_LINE;
     }
