+++ src/main/java/net/Mega2223/AguaEngine/graphics/objects/shapes/Oval.java
@@ -1,7 +1,7 @@
 package net.Mega2223.AguaEngine.graphics.objects.shapes;
 
 public class Oval extends Shape {
-    public float sizeX = 1;
-    public float sizeY = 1;
-    public Oval(float sizeX, float sizeY){this.sizeX = sizeX; this.sizeY = sizeY;type = SHAPE_OVAL;}
+    public double sizeX = 1;
+    public double sizeY = 1;
+    public Oval(double sizeX, double sizeY){this.sizeX = sizeX; this.sizeY = sizeY;type = SHAPE_OVAL;}
 }
