+++ src/main/java/net/Mega2223/AguaEngine/graphics/objects/shapes/Polygon.java
@@ -1,8 +1,8 @@
 package net.Mega2223.AguaEngine.graphics.objects.shapes;
 
 public class Polygon extends Shape {
-    public float[][] points = {};
-    public Polygon (float[][] points){
+    public double[][] points = {};
+    public Polygon (double[][] points){
         this.points = points;
         type = Shape.SHAPE_POLYGON;
     }
